move threading related function from doNd to threading.py
