+++ qcodes/utils/dataset/doNd.py
@@ -2,7 +2,6 @@ import logging
 import os
 import sys
 from abc import ABC, abstractmethod
-from collections import defaultdict
 from contextlib import contextmanager
 from typing import Callable, Dict, Iterator, List, Optional, Sequence, Tuple, Union
 
@@ -15,10 +14,10 @@ from qcodes.dataset.data_set_protocol import DataSetProtocol
 from qcodes.dataset.descriptions.detect_shapes import detect_shape_of_measurement
 from qcodes.dataset.descriptions.versioning.rundescribertypes import Shapes
 from qcodes.dataset.experiment_container import Experiment
-from qcodes.dataset.measurements import Measurement, res_type
+from qcodes.dataset.measurements import Measurement
 from qcodes.dataset.plotting import plot_dataset
-from qcodes.instrument.parameter import ParamDataType, _BaseParameter
-from qcodes.utils.threading import RespondingThread
+from qcodes.instrument.parameter import _BaseParameter
+from qcodes.utils.threading import process_params_meas
 
 ActionsT = Sequence[Callable[[], None]]
 
@@ -34,97 +33,11 @@ AxesTupleListWithDataSet = Tuple[
     List[Optional[matplotlib.colorbar.Colorbar]],
 ]
 
-OutType = List[res_type]
-
 LOG = logging.getLogger(__name__)
 
 
 class UnsafeThreadingException(Exception):
     pass
-
-
-class _ParamCaller:
-
-    def __init__(self, *parameters: _BaseParameter):
-
-        self._parameters = parameters
-
-    def __call__(self) -> Tuple[ParamDataType, ...]:
-        output = []
-        for param in self._parameters:
-            output.append(param.get())
-        return tuple(output)
-
-    def __repr__(self) -> str:
-        names = tuple(param.full_name for param in self._parameters)
-        return f"ParamCaller of {names}"
-
-
-def _instrument_to_param(
-        params: Sequence[ParamMeasT]
-) -> Dict[Optional[str], Tuple[_BaseParameter, ...]]:
-
-    real_parameters = [param for param in params
-                       if isinstance(param, _BaseParameter)]
-
-    output: Dict[Optional[str], Tuple[_BaseParameter, ...]] = defaultdict(tuple)
-    for param in real_parameters:
-        if param.root_instrument:
-            output[param.root_instrument.full_name] += (param,)
-        else:
-            output[None] += (param,)
-
-    return output
-
-
-def _call_params_threaded(param_meas: Sequence[ParamMeasT]) -> OutType:
-
-    inst_param_mapping = _instrument_to_param(param_meas)
-    executors = tuple(_ParamCaller(*param_list)
-                      for param_list in
-                      inst_param_mapping.values())
-
-    output: OutType = []
-    threads = [RespondingThread(target=executor)
-               for executor in executors]
-
-    for t in threads:
-        t.start()
-
-    for t, param_list in zip(threads, inst_param_mapping.values()):
-        thread_output = t.output()
-        assert thread_output is not None
-        for param, value in zip(param_list, thread_output):
-            output.append((param, value))
-
-    return output
-
-
-def _call_params(param_meas: Sequence[ParamMeasT]) -> OutType:
-
-    output: OutType = []
-
-    for parameter in param_meas:
-        if isinstance(parameter, _BaseParameter):
-            output.append((parameter, parameter.get()))
-        elif callable(parameter):
-            parameter()
-
-    return output
-
-
-def process_params_meas(
-    param_meas: Sequence[ParamMeasT],
-    use_threads: Optional[bool] = None
-) -> OutType:
-
-    if use_threads is None:
-        use_threads = config.dataset.use_threads
-
-    if use_threads:
-        return _call_params_threaded(param_meas)
-
-    return _call_params(param_meas)
 
 
 def _register_parameters(
